Delete HTTP webhook code
