+++ pkg/beyla/config_test.go
@@ -393,7 +393,6 @@ network:
 		Java:   obi.JavaConfig{Enabled: true, Timeout: 10 * time.Second},
 		Injector: SDKInject{
 			Webhook: WebhookConfig{
-				Enable:                false,
 				Port:                  8443,
 				Timeout:               30 * time.Second,
 				CertPath:              "/etc/webhook/certs/tls.crt",
@@ -774,7 +773,7 @@ func TestConfigRunsWithJustInjector(t *testing.T) {
 	userConfig := bytes.NewBufferString(`
 injector:
   webhook:
-    enable: true
+    external_deployment_name: foo/bar
     port: 8443
     cert_path: /etc/webhook/certs/tls.crt
     key_path: /etc/webhook/certs/tls.key
@@ -792,7 +791,7 @@ func TestConfigRunsWithJustInjectorButNotWithoutSDKPackage(t *testing.T) {
 	userConfig := bytes.NewBufferString(`
 injector:
   webhook:
-    enable: true
+    external_deployment_name: foo/bar
     port: 8443
     cert_path: /etc/webhook/certs/tls.crt
     key_path: /etc/webhook/certs/tls.key
@@ -809,7 +808,7 @@ func TestConfigRunsWithJustInjectorButNotWithoutHostPath(t *testing.T) {
 	userConfig := bytes.NewBufferString(`
 injector:
   webhook:
-    enable: true
+    external_deployment_name: foo/bar
     port: 8443
     cert_path: /etc/webhook/certs/tls.crt
     key_path: /etc/webhook/certs/tls.key
@@ -826,7 +825,7 @@ func TestConfigRunsWithJustInjectorButNotWithoutTraces(t *testing.T) {
 	userConfig := bytes.NewBufferString(`
 injector:
   webhook:
-    enable: true
+    external_deployment_name: foo/bar
     port: 8443
     cert_path: /etc/webhook/certs/tls.crt
     key_path: /etc/webhook/certs/tls.key
