--- pkg/webhook/example/apps.yaml
@@ -1,70 +1,0 @@
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: frontend
-  labels:
-    app: frontend
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: frontend
-  template:
-    metadata:
-      labels:
-        app: frontend
-    spec:
-      containers:
-      - name: frontend
-        image: ghcr.io/grafana/beyla/voxxed-days-crete/frontend:0.0.1
-        imagePullPolicy: IfNotPresent
-        ports:
-        - containerPort: 8080
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: frontend
-spec:
-  selector:
-    app: frontend
-  ports:
-    - protocol: TCP
-      port: 8080
-      targetPort: 8080
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: name-generator
-  labels:
-    app: name-generator
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: name-generator
-  template:
-    metadata:
-      labels:
-        app: name-generator
-    spec:
-      containers:
-      - name: name-generator
-        image: ghcr.io/grafana/beyla/voxxed-days-crete/name-generator:0.0.1
-        imagePullPolicy: IfNotPresent
-        ports:
-        - containerPort: 8081
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: name-generator
-spec:
-  selector:
-    app: name-generator
-  ports:
-    - protocol: TCP
-      port: 8081
-      targetPort: 8081
