+++ pkg/webhook/mutator_test.go
@@ -1,11 +1,7 @@
 package webhook
 
 import (
-	"bytes"
 	"log/slog"
-	"net/http"
-	"net/http/httptest"
-	"sync"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -15,7 +11,6 @@ import (
 
 	"go.opentelemetry.io/obi/pkg/appolly/app/svc"
 	"go.opentelemetry.io/obi/pkg/appolly/services"
-	"go.opentelemetry.io/obi/pkg/export/otel/otelcfg"
 	"go.opentelemetry.io/obi/pkg/kube/kubecache/informer"
 
 	"github.com/grafana/beyla/v3/pkg/beyla"
@@ -58,115 +53,6 @@ func TestErrorResponse(t *testing.T) {
 			assert.Equal(t, tt.expected, admResponse.Result.Message)
 		})
 	}
-}
-
-func TestPodMutator_HandleMutateRejectsOversizedBody(t *testing.T) {
-	mutator := &PodMutator{
-		logger: slog.With("component", "webhook.Mutator"),
-		cfg:    beyla.DefaultConfig(),
-	}
-
-	req := httptest.NewRequest(
-		http.MethodPost,
-		"/mutate",
-		bytes.NewReader(bytes.Repeat([]byte("a"), int(mutator.cfg.Injector.Webhook.MaxAdmissionBodySize.Value())+1)),
-	)
-	req.Header.Set("Content-Type", "application/json")
-
-	recorder := httptest.NewRecorder()
-	mutator.HandleMutate(recorder, req)
-
-	assert.Equal(t, http.StatusRequestEntityTooLarge, recorder.Code)
-	assert.Contains(t, recorder.Body.String(), "request body too large")
-}
-
-func TestPodMutator_HandleMutateRejectsWhenRequestLimitReached(t *testing.T) {
-	requestLimiter := make(chan struct{}, 1)
-	requestLimiter <- struct{}{}
-	mutator := &PodMutator{
-		logger:         slog.With("component", "webhook.Mutator"),
-		requestLimiter: requestLimiter,
-	}
-
-	req := httptest.NewRequest(
-		http.MethodPost,
-		"/mutate",
-		bytes.NewReader([]byte(`{}`)),
-	)
-	req.Header.Set("Content-Type", "application/json")
-
-	recorder := httptest.NewRecorder()
-	mutator.HandleMutate(recorder, req)
-
-	assert.Equal(t, http.StatusServiceUnavailable, recorder.Code)
-	assert.Contains(t, recorder.Body.String(), "too many concurrent mutation requests")
-}
-
-func TestPodMutator_HandleMutateAllows10ConcurrentRequests(t *testing.T) {
-	const concurrency = 10
-	requestLimiter := make(chan struct{}, concurrency)
-	mutator := &PodMutator{
-		logger:           slog.With("component", "webhook.Mutator"),
-		requestLimiter:   requestLimiter,
-		maxAdmissionSize: 3 * 1024 * 1024,
-	}
-
-	var wg sync.WaitGroup
-	codes := make([]int, concurrency)
-	start := make(chan struct{})
-
-	for i := 0; i < concurrency; i++ {
-		wg.Add(1)
-		go func(i int) {
-			defer wg.Done()
-			<-start
-			req := httptest.NewRequest(
-				http.MethodPost,
-				"/mutate",
-				bytes.NewReader([]byte(`{}`)),
-			)
-			req.Header.Set("Content-Type", "application/json")
-			recorder := httptest.NewRecorder()
-			mutator.HandleMutate(recorder, req)
-			codes[i] = recorder.Code
-		}(i)
-	}
-
-	close(start)
-	wg.Wait()
-
-	for i, code := range codes {
-		assert.NotEqual(t, http.StatusServiceUnavailable, code, "request %d was rejected by the limiter", i)
-	}
-}
-
-func TestPodMutator_HandleMutateDoesntRejectWhenRequestLimitNotReached(t *testing.T) {
-	requestLimiter := make(chan struct{}, 2)
-	requestLimiter <- struct{}{}
-	mutator := &PodMutator{
-		logger:           slog.With("component", "webhook.Mutator"),
-		requestLimiter:   requestLimiter,
-		maxAdmissionSize: 3 * 1024 * 1024,
-	}
-
-	req := httptest.NewRequest(
-		http.MethodPost,
-		"/mutate",
-		bytes.NewReader([]byte(`{}`)),
-	)
-	req.Header.Set("Content-Type", "application/json")
-
-	recorder := httptest.NewRecorder()
-	mutator.HandleMutate(recorder, req)
-
-	assert.Equal(t, http.StatusBadRequest, recorder.Code)
-	assert.NotContains(t, recorder.Body.String(), "too many concurrent mutation requests")
-
-	recorder = httptest.NewRecorder()
-	mutator.HandleMutate(recorder, req)
-
-	assert.Equal(t, http.StatusBadRequest, recorder.Code)
-	assert.NotContains(t, recorder.Body.String(), "too many concurrent mutation requests")
 }
 
 func TestPodMutator_CanInstrument(t *testing.T) {
@@ -496,401 +382,6 @@ func TestIsLDPreloadConflict(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			c := &corev1.Container{Env: tt.env}
 			assert.Equal(t, tt.expected, isLDPreloadConflict(c))
-		})
-	}
-}
-
-func TestPodMutator_MutatePod(t *testing.T) {
-	tests := []struct {
-		name           string
-		cfg            *beyla.Config
-		matcher        *PodMatcher
-		pod            *corev1.Pod
-		expectModified bool
-		expectLabel    bool
-		expectVolume   bool
-		expectEnvVars  bool
-	}{
-		{
-			name: "successful mutation - single container",
-			cfg: &beyla.Config{
-				Injector: beyla.SDKInject{
-					SDKPkgVersion:     "v0.0.3",
-					HostPathVolumeDir: "/var/lib/beyla/instrumentation",
-				},
-				Traces: otelcfg.TracesConfig{
-					CommonEndpoint: "http://localhost:4318",
-				},
-			},
-			matcher: &PodMatcher{
-				logger: slog.With("component", "webhook.Matcher"),
-				selectors: []services.Selector{
-					&services.GlobAttributes{
-						Metadata: map[string]*services.GlobAttr{
-							"k8s_namespace": strToGlob("*"),
-						},
-					},
-				},
-			},
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-pod",
-					Namespace: "default",
-				},
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{Name: "app"},
-					},
-				},
-			},
-			expectModified: true,
-			expectLabel:    true,
-			expectVolume:   true,
-			expectEnvVars:  true,
-		},
-		{
-			name: "already instrumented - skip mutation",
-			cfg: &beyla.Config{
-				Injector: beyla.SDKInject{
-					SDKPkgVersion:     "v0.0.3",
-					HostPathVolumeDir: "/var/lib/beyla/instrumentation",
-				},
-			},
-			matcher: &PodMatcher{
-				selectors: []services.Selector{
-					&services.GlobAttributes{
-						Metadata: map[string]*services.GlobAttr{
-							"k8s_namespace": strToGlob("*"),
-						},
-					},
-				},
-				logger: slog.With("component", "webhook.Matcher"),
-			},
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-pod",
-					Namespace: "default",
-					Labels: map[string]string{
-						instrumentedLabel: "v0.0.3",
-					},
-				},
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{Name: "app"},
-					},
-				},
-			},
-			expectModified: false,
-			expectLabel:    false,
-			expectVolume:   false,
-			expectEnvVars:  false,
-		},
-		{
-			name: "doesn't match selection - skip mutation",
-			cfg: &beyla.Config{
-				Injector: beyla.SDKInject{
-					SDKPkgVersion:     "v0.0.3",
-					HostPathVolumeDir: "/var/lib/beyla/instrumentation",
-				},
-			},
-			matcher: &PodMatcher{
-				logger: slog.With("component", "webhook.Matcher"),
-				selectors: []services.Selector{
-					&services.GlobAttributes{
-						Metadata: map[string]*services.GlobAttr{
-							"k8s_namespace": strToGlob("production"),
-						},
-					},
-				},
-			},
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-pod",
-					Namespace: "default", // doesn't match production
-				},
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{Name: "app"},
-					},
-				},
-			},
-			expectModified: false,
-			expectLabel:    false,
-			expectVolume:   false,
-			expectEnvVars:  false,
-		},
-		{
-			name: "successful mutation - multiple containers",
-			cfg: &beyla.Config{
-				Injector: beyla.SDKInject{
-					SDKPkgVersion:     "v0.0.3",
-					HostPathVolumeDir: "/var/lib/beyla/instrumentation",
-				},
-				Traces: otelcfg.TracesConfig{
-					CommonEndpoint: "http://localhost:4318",
-				},
-			},
-			matcher: &PodMatcher{
-				logger: slog.With("component", "webhook.Matcher"),
-				selectors: []services.Selector{
-					&services.GlobAttributes{
-						Metadata: map[string]*services.GlobAttr{
-							"k8s_namespace": strToGlob("*"),
-						},
-					},
-				},
-			},
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-pod",
-					Namespace: "default",
-				},
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{Name: "app1"},
-						{Name: "app2"},
-						{Name: "app3"},
-					},
-				},
-			},
-			expectModified: true,
-			expectLabel:    true,
-			expectVolume:   true,
-			expectEnvVars:  true,
-		},
-		{
-			name: "all containers have empty LD_PRELOAD - should instrument",
-			cfg: &beyla.Config{
-				Injector: beyla.SDKInject{
-					SDKPkgVersion:     "v0.0.3",
-					HostPathVolumeDir: "/var/lib/beyla/instrumentation",
-				},
-				Traces: otelcfg.TracesConfig{
-					CommonEndpoint: "http://localhost:4318",
-				},
-			},
-			matcher: &PodMatcher{
-				logger: slog.With("component", "webhook.Matcher"),
-				selectors: []services.Selector{
-					&services.GlobAttributes{
-						Metadata: map[string]*services.GlobAttr{
-							"k8s_namespace": strToGlob("*"),
-						},
-					},
-				},
-			},
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-pod",
-					Namespace: "default",
-				},
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{
-							Name: "app",
-							Env: []corev1.EnvVar{
-								{Name: envVarLdPreloadName, Value: ""},
-							},
-						},
-					},
-				},
-			},
-			expectModified: true,
-			expectLabel:    true,
-			expectVolume:   true,
-			expectEnvVars:  true,
-		},
-		{
-			name: "all containers have genuine LD_PRELOAD conflict - should skip",
-			cfg: &beyla.Config{
-				Injector: beyla.SDKInject{
-					SDKPkgVersion:     "v0.0.3",
-					HostPathVolumeDir: "/var/lib/beyla/instrumentation",
-				},
-			},
-			matcher: &PodMatcher{
-				logger: slog.With("component", "webhook.Matcher"),
-				selectors: []services.Selector{
-					&services.GlobAttributes{
-						Metadata: map[string]*services.GlobAttr{
-							"k8s_namespace": strToGlob("*"),
-						},
-					},
-				},
-			},
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-pod",
-					Namespace: "default",
-				},
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{
-							Name: "app1",
-							Env: []corev1.EnvVar{
-								{Name: envVarLdPreloadName, Value: "/some/other/lib.so"},
-							},
-						},
-						{
-							Name: "app2",
-							Env: []corev1.EnvVar{
-								{Name: envVarLdPreloadName, Value: "/another/lib.so"},
-							},
-						},
-					},
-				},
-			},
-			expectModified: false,
-			expectLabel:    false,
-			expectVolume:   false,
-			expectEnvVars:  false,
-		},
-		{
-			name: "container with Beyla own LD_PRELOAD value - not treated as conflict",
-			cfg: &beyla.Config{
-				Injector: beyla.SDKInject{
-					SDKPkgVersion:     "v0.0.3",
-					HostPathVolumeDir: "/var/lib/beyla/instrumentation",
-				},
-				Traces: otelcfg.TracesConfig{
-					CommonEndpoint: "http://localhost:4318",
-				},
-			},
-			matcher: &PodMatcher{
-				logger: slog.With("component", "webhook.Matcher"),
-				selectors: []services.Selector{
-					&services.GlobAttributes{
-						Metadata: map[string]*services.GlobAttr{
-							"k8s_namespace": strToGlob("*"),
-						},
-					},
-				},
-			},
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-pod",
-					Namespace: "default",
-				},
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{
-							Name: "app-with-beyla-preload",
-							Env: []corev1.EnvVar{
-								{Name: envVarLdPreloadName, Value: envVarLdPreloadValue},
-							},
-						},
-						{Name: "app-clean"},
-					},
-				},
-			},
-			expectModified: true,
-			expectLabel:    true,
-			expectVolume:   true,
-			expectEnvVars:  true,
-		},
-		{
-			name: "skip container with existing LD_PRELOAD",
-			cfg: &beyla.Config{
-				Injector: beyla.SDKInject{
-					SDKPkgVersion:     "v0.0.3",
-					HostPathVolumeDir: "/var/lib/beyla/instrumentation",
-				},
-				Traces: otelcfg.TracesConfig{
-					CommonEndpoint: "http://localhost:4318",
-				},
-			},
-			matcher: &PodMatcher{
-				logger: slog.With("component", "webhook.Matcher"),
-				selectors: []services.Selector{
-					&services.GlobAttributes{
-						Metadata: map[string]*services.GlobAttr{
-							"k8s_namespace": strToGlob("*"),
-						},
-					},
-				},
-			},
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-pod",
-					Namespace: "default",
-				},
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{
-							Name: "app-with-preload",
-							Env: []corev1.EnvVar{
-								{Name: envVarLdPreloadName, Value: "/some/other/lib.so"},
-							},
-						},
-						{Name: "app-without-preload"},
-					},
-				},
-			},
-			expectModified: true,
-			expectLabel:    true,
-			expectVolume:   true,
-			expectEnvVars:  true, // at least one container should be instrumented
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			mutator := &PodMutator{
-				logger:   slog.With("component", "webhook.Mutator"),
-				cfg:      tt.cfg,
-				matcher:  tt.matcher,
-				endpoint: "http://localhost:4318",
-				proto:    "http/protobuf",
-			}
-
-			selector, matched := mutator.matchesSelection(&tt.pod.ObjectMeta)
-			var modified bool
-			if matched {
-				modified, _ = mutator.mutatePod(tt.pod, selector)
-			}
-
-			assert.Equal(t, tt.expectModified, modified, "mutation result mismatch")
-
-			if tt.expectLabel {
-				label, ok := tt.pod.Labels[instrumentedLabel]
-				assert.True(t, ok, "instrumented label should be present")
-				assert.Equal(t, tt.cfg.Injector.SDKPkgVersion, label)
-			}
-
-			if tt.expectVolume {
-				found := false
-				for _, vol := range tt.pod.Spec.Volumes {
-					if vol.Name == injectVolumeName {
-						found = true
-						assert.NotNil(t, vol.HostPath)
-						assert.Equal(t, tt.cfg.Injector.HostPathVolumeDir+"/"+tt.cfg.Injector.SDKPkgVersion, vol.HostPath.Path)
-						break
-					}
-				}
-				assert.True(t, found, "inject volume should be present")
-			}
-
-			if tt.expectEnvVars {
-				// Check at least one container has instrumentation env vars
-				foundInstrumented := false
-				for _, c := range tt.pod.Spec.Containers {
-					hasLdPreload := false
-					hasConfigFile := false
-					for _, env := range c.Env {
-						if env.Name == envVarLdPreloadName && env.Value == envVarLdPreloadValue {
-							hasLdPreload = true
-						}
-						if env.Name == envOtelInjectorConfigFileName {
-							hasConfigFile = true
-						}
-					}
-					if hasLdPreload && hasConfigFile {
-						foundInstrumented = true
-						break
-					}
-				}
-				assert.True(t, foundInstrumented, "at least one container should be instrumented")
-			}
 		})
 	}
 }
@@ -1237,101 +728,6 @@ func TestSetEnvVar(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			setEnvVar(tt.container, tt.envVarName, tt.value)
 			assert.Equal(t, tt.expectedEnv, tt.container.Env)
-		})
-	}
-}
-
-func TestPodMutator_AddEnvVars(t *testing.T) {
-	tests := []struct {
-		name             string
-		meta             *metav1.ObjectMeta
-		container        *corev1.Container
-		cfg              *beyla.Config
-		endpoint         string
-		proto            string
-		exportHeaders    map[string]string
-		expectedEnvCount int
-		checkEnvVars     map[string]string
-	}{
-		{
-			name: "add basic env vars",
-			meta: &metav1.ObjectMeta{
-				Name:      "test-pod",
-				Namespace: "default",
-			},
-			container: &corev1.Container{
-				Name: "test-container",
-			},
-			cfg: &beyla.Config{
-				Injector: beyla.SDKInject{
-					SDKPkgVersion: "v0.0.3",
-				},
-			},
-			endpoint:      "http://localhost:4318",
-			proto:         "http/protobuf",
-			exportHeaders: map[string]string{},
-			checkEnvVars: map[string]string{
-				envVarSDKVersion:                "v0.0.3",
-				envVarLdPreloadName:             envVarLdPreloadValue,
-				envOtelInjectorConfigFileName:   envOtelInjectorConfigFileValue,
-				envOtelExporterOtlpEndpointName: "http://localhost:4318",
-				envOtelExporterOtlpProtocolName: "http/protobuf",
-				envOtelSemConvStabilityName:     "http",
-			},
-		},
-		{
-			name: "add env vars with export headers",
-			meta: &metav1.ObjectMeta{
-				Name:      "test-pod",
-				Namespace: "default",
-			},
-			container: &corev1.Container{
-				Name: "test-container",
-			},
-			cfg: &beyla.Config{
-				Injector: beyla.SDKInject{
-					SDKPkgVersion: "v0.0.4",
-				},
-			},
-			endpoint: "http://localhost:4318",
-			proto:    "grpc",
-			exportHeaders: map[string]string{
-				"OTEL_EXPORTER_OTLP_HEADERS": "Authorization=Bearer token",
-			},
-			checkEnvVars: map[string]string{
-				envVarSDKVersion:                "v0.0.4",
-				envVarLdPreloadName:             envVarLdPreloadValue,
-				envOtelExporterOtlpEndpointName: "http://localhost:4318",
-				envOtelExporterOtlpProtocolName: "grpc",
-				"OTEL_EXPORTER_OTLP_HEADERS":    "Authorization=Bearer token",
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			mutator := &PodMutator{
-				logger:        slog.With("component", "webhook.test"),
-				cfg:           tt.cfg,
-				endpoint:      tt.endpoint,
-				proto:         tt.proto,
-				exportHeaders: tt.exportHeaders,
-			}
-
-			mutator.addEnvVars(tt.meta, tt.container, nil)
-
-			// Check that expected env vars are present
-			for key, expectedValue := range tt.checkEnvVars {
-				found := false
-				for _, env := range tt.container.Env {
-					if env.Name == key {
-						found = true
-						assert.Equal(t, expectedValue, env.Value, "env var %s has wrong value", key)
-						break
-					}
-				}
-				assert.True(t, found, "env var %s not found", key)
-			}
 		})
 	}
 }
